third_party/WebKit/LayoutTests/bluetooth/server/getPrimaryService/gen-no-permission-for-any-service.html

<!-- Generated by //third_party/WebKit/LayoutTests/bluetooth/generate.py -->
<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/bluetooth/bluetooth-helpers.js"></script>
<script src="../../../resources/bluetooth/web-bluetooth-test.js"></script>
<script src="../../../resources/mojo-helpers.js"></script>
<script>
'use strict';
promise_test(() => {
  return getHealthThermometerDevice({acceptAllDevices: true})
    .then(([device]) => assert_promise_rejects_with_message(
        device.gatt.getPrimaryService('heart_rate'),
        new DOMException('Origin is not allowed to access any service. Tip: ' +
                         'Add the service UUID to \'optionalServices\' in ' +
                         'requestDevice() options. https://goo.gl/HxfxSQ',
                         'SecurityError')));
}, 'Request for present service without permission to access any service. ' +
   'Reject with SecurityError.');

</script>
